salary_predictor/requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.3.0 # pyup: != 11.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==25.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.9.0  # https://github.com/evansd/whitenoise
redis==6.2.0  # https://github.com/redis/redis-py
hiredis==3.2.1  # https://github.com/redis/hiredis-py
celery==5.5.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.8.1  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==5.1.11  # pyup: < 5.2 # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.10.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.4  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2025.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==6.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.16.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.7.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.28.0  # https://github.com/tfranzel/drf-spectacular
django-webpack-loader==3.2.1  # https://github.com/django-webpack/django-webpack-loader
django-widget-tweaks==1.5.0
scikit-learn==1.7.1
joblib==1.5.1
pandas==2.3.1
numpy==2.2.0
xgboost==3.0.2
matplotlib==3.10.3
seaborn==0.13.2
shap==0.48.0
category_encoders==2.8.1
lightgbm==4.6.0
